% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map_dbl_hack.R
\name{map_dbl_hack}
\alias{map_dbl_hack}
\title{Map given function, handle null as NA_real_ and flatten_dbl()}
\usage{
map_dbl_hack(.x, .f, ...)
}
\arguments{
\item{.x}{An interable data object}

\item{.f}{A function to map over the data}

\item{...}{Extra arguments to `map()`}
}
\description{
Map given function, handle null as NA_real_ and flatten_dbl()
}
\details{
Use this function to extract elements with NAs from a list of lists
into a tibble.
}
\examples{
#data(animal_encounters)
#nn <- map_chr_hack(animal_encounters$content, c("properties", "name"))
#testthat::expect_true(is.na(nn[[1]]))
#testthat::expect_equal(nn[[2]], animal_encounters$content[[2]]$properties$name)
}
\author{
Jennifer Bryan https://github.com/jennybc/
}

