% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/biosys_get.R
\name{print.biosys_api_response}
\alias{print.biosys_api_response}
\title{S3 print method for 'biosys_api_response'.}
\usage{
\method{print}{biosys_api_response}(x, ...)
}
\arguments{
\item{x}{An object of class `biosys_api_response` as returned by
\code{\link{get_biosys}}.}

\item{...}{Extra parameters for `print`}
}
\description{
Prints a short representation of data returned by \
  code{\link{get_biosys}}.
}

